Rename mixins module to tornado.
Tried to make it clear that tornado is an optional requirement.
