+++ setup.cfg
@@ -29,9 +29,10 @@ classifiers =
 [options]
 packages = find:
 install_requires =
-    tornado>=5
 
 [options.extras_require]
+tornado =
+    tornado>=5
 dev =
     asynctest==0.13.0
     coverage==5.5
@@ -39,6 +40,7 @@ dev =
     flake8-import-order==0.18.1
     sphinx==3.5.2
     sphinx-autodoc-typehints==1.11.1
+    tornado>=5
     yapf==0.30.0
 
 [options.packages.find]
