+++ tox.ini
@@ -4,7 +4,7 @@ toxworkdir = ./build/tox
 
 [testenv]
 deps =
-	.[dev]
+	.[dev,tornado]
 commands =
 	python -m unittest
